@@ -1,4 +1,4 @@
-class AddColumnAvatarToUsers < ActiveRecord::Migration
+class AddColumnAvatarToUsers < ActiveRecord::Migration[5.1]
   def change
     add_column :users, :avatar, :string
   end
