@@ -1,6 +1,6 @@
 # This migration comes from acts_as_taggable_on_engine (originally 1)
-class ActsAsTaggableOnMigration < ActiveRecord::Migration
-  def self.up
+class ActsAsTaggableOnMigration < ActiveRecord::Migration[5.1]
+  def up
     create_table :tags do |t|
       t.string :name
     end
@@ -20,11 +20,10 @@ class ActsAsTaggableOnMigration < ActiveRecord::Migration
       t.datetime :created_at
     end
 
-    add_index :taggings, :tag_id
     add_index :taggings, [:taggable_id, :taggable_type, :context]
   end
 
-  def self.down
+  def down
     drop_table :taggings
     drop_table :tags
   end
