@@ -1,4 +1,4 @@
-class AddColumnDeletedAtToCategories < ActiveRecord::Migration
+class AddColumnDeletedAtToCategories < ActiveRecord::Migration[5.1]
   def change
     add_column :categories, :deleted_at, :datetime
   end
