@@ -1,4 +1,4 @@
-class AddObjectChangesToVersions < ActiveRecord::Migration
+class AddObjectChangesToVersions < ActiveRecord::Migration[5.1]
 
   # The largest text column available in all supported RDBMS.
   # See `create_versions.rb` for details.
