@@ -1,4 +1,4 @@
-class CreateAuthorizations < ActiveRecord::Migration
+class CreateAuthorizations < ActiveRecord::Migration[5.1]
   def change
     create_table :authorizations do |t|
       t.integer :provider
