@@ -10,9 +10,13 @@
 #  sort       :integer
 #
 
-FactoryGirl.define do
+FactoryBot.define do
   factory :category do
     sequence(:name) { |n| "category #{n}" }
+
+    trait :admin_creation do
+      tag_list { ['tag1', 'tag2', 'tag3'] }
+    end
   end
 
 end
