@@ -7,7 +7,7 @@
 #  taggings_count :integer          default(0)
 #
 
-FactoryGirl.define do
+FactoryBot.define do
   factory :tag do
 
   end
