@@ -18,29 +18,33 @@
 #  confirmed_at           :datetime
 #  confirmation_sent_at   :datetime
 #  unconfirmed_email      :string
-#  created_at             :datetime
-#  updated_at             :datetime
+#  created_at             :datetime         not null
+#  updated_at             :datetime         not null
 #  admin                  :boolean          default(FALSE)
 #  avatar                 :string
 #
 
-FactoryGirl.define do
+FactoryBot.define do
   factory :user do
     name '5Fpro'
     sequence(:email) { |n| "user#{n}@5fpro.com" }
     password '12341234'
     confirmed_at Time.now
 
-    factory :admin_user do
+    trait :admin do
       admin true
     end
 
-    factory :unconfirmed_user do
+    trait :unconfirmed do
       confirmed_at nil
     end
 
-    factory :user_with_avatar do
+    trait :with_avatar do
       avatar { File.open(Rails.root.join('spec', 'fixtures', '5fpro.png')) }
+    end
+
+    trait :admin_creation do
+      confirmed_at nil
     end
   end
 end
