@@ -52,13 +52,13 @@ RSpec.configure do |config|
   config.infer_spec_type_from_file_location!
   config.include Util
   config.include Webmock
-  config.include DataMaker
   config.include RequestClient, type: :request
-  config.include HtmlMatchers, type: :request
-  config.include FactoryGirl::Syntax::Methods
+  config.include FactoryBot::Syntax::Methods
+  config.include Devise::Test::ControllerHelpers, type: :controller
 
   config.before(:each){ webmock_all! }
   config.before(:each){ Redis.current.flushdb }
+  config.before(:each){ Rails.cache.clear }
 
   config.after { Timecop.return }
   # uncomment if you need specific time zone in default
