@@ -1,13 +1,13 @@
 class BaseContext
-  extend ActiveModel::Callbacks
-  define_model_callbacks :perform
-
   include Rails.application.routes.url_helpers
-  include ErrorHandler
+  include ObjectErrorsConcern
+
+  define_model_callbacks :perform
 
   private
 
   def permit_params(params, *cols)
-    ActionController::Parameters.new(params).permit(cols)
+    params = params.permit! if params.respond_to?(:permit!)
+    ActionController::Parameters.new(params.to_h.with_indifferent_access).permit(cols)
   end
 end
