@@ -4,14 +4,18 @@ module Api
     before_action :accept_cors
     before_action :set_default_format
 
-    include ::Api::ErrorResponseHandler
+    include ErrorResponseConcern
 
     def index
       render json: { ok: true, params: params.permit(params.keys).to_h }
     end
 
     def error
-      Error.raise!(params.require(:error), messages: 'Error object test', data: params.permit(params.keys).to_h)
+      raise Api::ClientAuthError
+    end
+
+    def respond_404
+      raise RoutingError.new(request: request)
     end
 
     private
