@@ -1,6 +1,5 @@
 class BaseController < ApplicationController
+  before_action :set_meta
 
-  def index
-    set_meta(title: ENV['APP_NAME'])
-  end
+  def index; end
 end
