@@ -4,13 +4,16 @@ class PicPreviewInput < SimpleForm::Inputs::FileInput
     # :version is a custom attribute from :input_html hash, so you can pick custom sizes
     version = input_html_options.delete(:version)
     out = ActiveSupport::SafeBuffer.new
+    input_html_classes.push('form-control')
     out << @builder.file_field(attribute_name, input_html_options)
     if object.send("#{attribute_name}?")
-      remove_chk = @builder.check_box("remove_#{attribute_name}") # carrierwave: remove_xxx
+      remove_chk = @builder.check_box("remove_#{attribute_name}", class: 'flat') # carrierwave: remove_xxx
       out << '<br />'.html_safe
-      out << "<label>#{remove_chk} remove</label>".html_safe # TRANSLATION: I18n here
+      remove_text = I18n.t('simple_form.buttons.remove', default: 'Remove')
+      out << "<label>#{remove_chk} #{remove_text}</label>".html_safe
       out << '<br />'.html_safe
-      out << template.image_tag(object.send(attribute_name).tap { |o| break o.send(version) if version }.send('url'))
+      img_path = object.send(attribute_name).tap { |o| break o.send(version) if version }.send('url')
+      out << template.image_tag(img_path)
     end
     out
   end
