@@ -1,5 +1,5 @@
 require 'dotenv'
-Dotenv.load
+Dotenv.overload
 
 # see more
 #   https://github.com/tablexi/capistrano3-unicorn/blob/master/examples/unicorn.rb
@@ -8,14 +8,16 @@ working_directory "#{app_path}/current"
 pid "#{app_path}/current/tmp/pids/unicorn.pid"
 listen "/tmp/unicorn.#{ENV['APP_NAME']}.sock"
 
+# default would be stderr_path
 # log rotate config example
 #   https://github.com/defunkt/unicorn/blob/master/examples/logrotate.conf
-stderr_path 'log/unicorn.error.log'
-stdout_path 'log/unicorn.log'
+stderr_path "#{app_path}/current/log/unicorn.log"
+stdout_path "#{app_path}/current/log/unicorn.log"
 
 worker_processes 1
 
 before_exec do |server|
+  Dotenv.overload
   ENV['BUNDLE_GEMFILE'] = "#{app_path}/current/Gemfile"
 end
 
